ai_recommendation_service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.7.0
numpy==1.26.4
scikit-learn==1.5.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
